requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pytesseract==0.3.10
aiosqlite==0.19.0
asyncio==3.4.3
python-switchbot==2.3.1
google-genai